# Change fields to properties in AST public API