--- a/src/Esprima/Ast/AwaitExpression.cs
+++ b/src/Esprima/Ast/AwaitExpression.cs
@@ -1,15 +1,16 @@
-﻿using Esprima.Utils;
+﻿using System.Runtime.CompilerServices;
+using Esprima.Utils;
 
 namespace Esprima.Ast
 {
     public sealed class AwaitExpression : Expression
     {
-        public readonly Expression Argument;
-
         public AwaitExpression(Expression argument) : base(Nodes.AwaitExpression)
         {
             Argument = argument;
         }
+
+        public Expression Argument { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
         public override NodeCollection ChildNodes => new(Argument);