--- a/src/Esprima/Ast/BigIntLiteral.cs
+++ b/src/Esprima/Ast/BigIntLiteral.cs
@@ -1,16 +1,16 @@
 ﻿using System.Numerics;
+using System.Runtime.CompilerServices;
 
 namespace Esprima.Ast
 {
     public sealed class BigIntLiteral : Literal
     {
-        public readonly string BigInt;
-
-        public new BigInteger? BigIntValue => (BigInteger?) Value;
-
         public BigIntLiteral(BigInteger value, string raw) : base(TokenType.BigIntLiteral, value, raw)
         {
-            BigInt = raw;
         }
+
+        public string BigInt { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => Raw; }
+
+        public new BigInteger? BigIntValue { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => (BigInteger?) Value; }
     }
 }