--- a/src/Esprima/Ast/CallExpression.cs
+++ b/src/Esprima/Ast/CallExpression.cs
@@ -1,13 +1,11 @@
-﻿using Esprima.Utils;
+﻿using System.Runtime.CompilerServices;
+using Esprima.Utils;
 
 namespace Esprima.Ast
 {
     public sealed class CallExpression : Expression
     {
         private readonly NodeList<Expression> _arguments;
-
-        public readonly Expression Callee;
-        public readonly bool Optional;
 
         public CallExpression(
             Expression callee,
@@ -19,9 +17,11 @@ namespace Esprima.Ast
             Optional = optional;
         }
 
-        public ref readonly NodeList<Expression> Arguments => ref _arguments;
+        public Expression Callee { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
+        public ref readonly NodeList<Expression> Arguments { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => ref _arguments; }
+        public bool Optional { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
-        public override NodeCollection ChildNodes => GenericChildNodeYield.Yield(Callee, _arguments);
+        public override NodeCollection ChildNodes => GenericChildNodeYield.Yield(Callee, Arguments);
 
         protected internal override object? Accept(AstVisitor visitor)
         {