--- a/src/Esprima/Ast/Directive.cs
+++ b/src/Esprima/Ast/Directive.cs
@@ -1,13 +1,15 @@
-﻿namespace Esprima.Ast
+﻿using System.Runtime.CompilerServices;
+
+namespace Esprima.Ast
 {
     public sealed class Directive : ExpressionStatement
     {
-        public readonly string Directiv;
-
         public Directive(Expression expression, string directive) : base(expression)
         {
             Directiv = directive;
         }
+
+        public string Directiv { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; }
 
         protected override ExpressionStatement Rewrite(Expression expression)
         {