--- a/src/Esprima/Ast/EmptyStatement.cs
+++ b/src/Esprima/Ast/EmptyStatement.cs
@@ -9,7 +9,7 @@ namespace Esprima.Ast
         }
 
         public override NodeCollection ChildNodes => NodeCollection.Empty;
-        
+
         protected internal override object? Accept(AstVisitor visitor)
         {
             return visitor.VisitEmptyStatement(this);