… src/Esprima/Ast/ExportDefaultDeclaration.cs
… src/Esprima/Ast/ExportNamedDeclaration.cs
… src/Esprima/Ast/ExportSpecifier.cs
… src/Esprima/Ast/ExpressionStatement.cs
… src/Esprima/Ast/ForInStatement.cs
… src/Esprima/Ast/ForOfStatement.cs
… src/Esprima/Ast/ForStatement.cs
… src/Esprima/Ast/FunctionDeclaration.cs
… src/Esprima/Ast/FunctionExpression.cs
… src/Esprima/Ast/Identifier.cs
… src/Esprima/Ast/IfStatement.cs
… src/Esprima/Ast/Import.cs
… src/Esprima/Ast/ImportAttribute.cs
… src/Esprima/Ast/ImportDeclaration.cs
… src/Esprima/Ast/ImportDeclarationSpecifier.cs
… src/Esprima/Ast/ImportSpecifier.cs
… src/Esprima/Ast/Jsx/JsxAttribute.cs
… src/Esprima/Ast/Jsx/JsxClosingElement.cs
… src/Esprima/Ast/Jsx/JsxElement.cs
… src/Esprima/Ast/Jsx/JsxExpression.cs
… src/Esprima/Ast/Jsx/JsxExpressionContainer.cs
… src/Esprima/Ast/Jsx/JsxIdentifier.cs
… src/Esprima/Ast/Jsx/JsxMemberExpression.cs
… src/Esprima/Ast/Jsx/JsxNamespacedName.cs
… src/Esprima/Ast/Jsx/JsxOpeningElement.cs
… src/Esprima/Ast/Jsx/JsxOpeningFragment.cs
… src/Esprima/Ast/Jsx/JsxSpreadAttribute.cs
… src/Esprima/Ast/Jsx/JsxText.cs
… src/Esprima/Ast/LabeledStatement.cs
… src/Esprima/Ast/Literal.cs