This PR addresses https://github.com/sebastienros/esprima-dotnet/issues/264 and, partially, https://github.com/sebastienros/esprima-dotnet/issues/66

Phew, this was some epic sh*t. It took all my text editing skills... 😄 But according to the benchmark it was worth it:

Before:

|  Method |        Job |              Runtime |    Toolchain |     Mean |    Error |   StdDev | Ratio | Rank | Allocated |
|-------- |----------- |--------------------- |------------- |---------:|---------:|---------:|------:|-----:|----------:|
|   Visit | Job-ZDQNTH |             .NET 6.0 |        net60 | 15.18 ms | 0.031 ms | 0.026 ms |  0.97 |    1 |      16 B |
|   Visit | Job-IEZQPH |        .NET Core 3.1 | netcoreapp31 | 15.24 ms | 0.029 ms | 0.027 ms |  0.98 |    1 |         - |
|   Visit | Job-WZKQTX | .NET Framework 4.6.2 |       net462 | 15.63 ms | 0.014 ms | 0.012 ms |  1.00 |    2 |         - |
|         |            |                      |              |          |          |          |       |      |           |
| Rewrite | Job-ZDQNTH |             .NET 6.0 |        net60 | 37.70 ms | 0.134 ms | 0.125 ms |  0.75 |    1 |      75 B |
| Rewrite | Job-IEZQPH |        .NET Core 3.1 | netcoreapp31 | 40.52 ms | 0.061 ms | 0.054 ms |  0.81 |    2 |         - |
| Rewrite | Job-WZKQTX | .NET Framework 4.6.2 |       net462 | 50.06 ms | 0.298 ms | 0.279 ms |  1.00 |    3 |         - |

After:

|  Method |        Job |              Runtime |    Toolchain |     Mean |    Error |   StdDev | Ratio | Rank | Allocated |
|-------- |----------- |--------------------- |------------- |---------:|---------:|---------:|------:|-----:|----------:|
|   Visit | Job-PALKNJ |             .NET 6.0 |        net60 | 15.26 ms | 0.046 ms | 0.043 ms |  0.98 |    2 |      16 B |
|   Visit | Job-SJJCSG |        .NET Core 3.1 | netcoreapp31 | 14.96 ms | 0.019 ms | 0.017 ms |  0.96 |    1 |         - |
|   Visit | Job-QAMQPG | .NET Framework 4.6.2 |       net462 | 15.55 ms | 0.016 ms | 0.013 ms |  1.00 |    3 |         - |
|         |            |                      |              |          |          |          |       |      |           |
| Rewrite | Job-PALKNJ |             .NET 6.0 |        net60 | 36.63 ms | 0.190 ms | 0.178 ms |  0.72 |    1 |      75 B |
| Rewrite | Job-SJJCSG |        .NET Core 3.1 | netcoreapp31 | 43.10 ms | 0.082 ms | 0.073 ms |  0.85 |    2 |         - |
| Rewrite | Job-QAMQPG | .NET Framework 4.6.2 |       net462 | 50.52 ms | 0.119 ms | 0.111 ms |  1.00 |    3 |         - |

So it looks like aggressive inlining did its job. (I didn't measure parsing because construction of the tree remained the same as I kept the fields for the constructors and other internal parts.) However, to be 100% certain, it'd be great if someone repeated the benchmarks and verified these result.

With these changes, the AST is now almost fully immutable. It's almost because I haven't touched `Node.Location` and `Node.Range`. These are still public fields and I'm somewhat puzzled as to what to do with them. The crux of the problem is that these properties aren't included in the constructors but should be settable also by end users. Including them in the ctors would be a massive BC, so that's absolutely out of question. Our options are the following:
1. We can simply wrap them in properties. E.g.
    ```csharp
    public Range Range { [MethodImpl(MethodImplOptions.AggressiveInlining)] get; [MethodImpl(MethodImplOptions.AggressiveInlining)] init; }
    ```
    However, this is certainly worse performance-wise than we have currently (i.e. simple fields) because `Range` and `Location` are value types, so there would be inevitable copying.
2. We can return by-ref but then the properties can't have a setter. We could provide some `SetRange` and `SetLocation` methods but then it's not immutable again. For init-only set we need a property. Thus, if we want by-ref return we'd need two properties:
    ```csharp
    public ref readonly Range RangeRef { [MethodImpl(MethodImplOptions.AggressiveInlining)] get => ref _range; }
    public Range Range { get => _range; init => _range = value; }
    ```
    This is ugly but the only way I'm aware of to retain perf. and make the API immutable for consumers at the same time.
3. We leave them alone and have a semi-immutable API.

Which one do you prefer?