@lahma 2022-06-10
I can confirm that there is no noticeable difference, run the results for main and this using net6.0:

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=7.0.100-preview.2.22153.17
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
  DefaultJob : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT


```

## Esprima.Benchmark.VisitorBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Visit|12.61 ms|0.124 ms|20 B|
| **New** |	| **12.38 ms (-2%)** | **0.069 ms** | **20 B (0%)** |
| Old |Rewrite|23.20 ms|0.186 ms|33 B|
| **New** |	| **22.64 ms (-2%)** | **0.257 ms** | **33 B (0%)** |


> Which one do you prefer?

Copying `Range` *shouldn't* be a big problem as it's a small struct. `Location` in the other hand is somewhat bigger already and might cause some perf hit. Maybe something that needs to be measured again. This time I think the benchmark would be parsing speed though.

We can leave that to another PR/later stage too though. So option 1 or 3 I guess.