@adams85 2022-06-10
> Copying Range shouldn't be a big problem as it's a small struct. Location in the other hand is somewhat bigger already and might cause some perf hit. Maybe something that needs to be measured again. This time I think the benchmark would be parsing speed though.

> We can leave that to another PR/later stage too though. So option 1 or 3 I guess.

Then I propose option 1 with a twist: let's change the fields to internal so we can use those internally for the sake of max. perf and provide the "slow" (copying) properties for end users. See the 2. commit.

If this is acceptable than the PR is ready.