@lahma 2022-06-10
Noticed one small problem with latest code on Jint side:

https://github.com/sebastienros/jint/blob/3bc93cf385795ad06af98376a99b8ed17cd2b615/Jint/Native/Json/JsonParser.cs#L522-L541

Probably needs investigation why this assignment is done or needed, I'm not familiar with this part.