@adams85 2022-06-10
> Excellent work once again!

Thank you!

> Noticed one small problem with latest code on Jint side:

Maybe add an `InternalsVisibleTo` for Jint? Or just scrap the idea of init-only properties for `Range` and `Location` and revert them to public fields?