@lahma 2022-06-10
InternalsVisible could work. Would be nice to know what other potentially breaking consumers there are, ones wanting to update this info. But as said, not a critical thing from Jint's side at the moment.